+++ requirements.txt
@@ -1,10 +1,9 @@
 # pip install -r requirements.txt
-supervisely==6.7.21
-supervisely[apps]==6.7.21
+supervisely[apps]==6.18.1
+
 # base ----------------------------------------
 PyYAML==5.4.1
 tensorboard==2.4.1
-tqdm==4.56.0
 
 # plotting ------------------------------------
 seaborn==0.11.1
@@ -12,7 +11,7 @@ seaborn==0.11.1
 # export --------------------------------------
 coremltools==4.1
 onnx>=1.8.0
-scikit-learn==0.19.2  # for coreml quantization
+scikit-learn==1.1.1  # for coreml quantization
 
 # extras --------------------------------------
 thop==0.0.31-2005241907  # FLOPS computation
