+++ supervisely_app/export_weights/config.json
