+++ supervisely_app/export_weights/debug.env
