+++ supervisely_app/export_weights/src/modal.html
