+++ supervisely_app/serve/config.json
