+++ supervisely_app/serve/custom_settings.yaml
