+++ supervisely_app/serve/debug.env
@@ -4,17 +4,17 @@ DEBUG_APP_DIR="/app_debug_data"
 DEBUG_CACHE_DIR="/app_cache"
 LOG_LEVEL="trace"
 
-TASK_ID=4585
+TASK_ID=16752
 
-context.teamId=237
-context.workspaceId=311
+context.teamId=312
+context.workspaceId=455
 
-modal.state.modelWeightsOptions="custom"
+modal.state.modelWeightsOptions="pretrained"
 #"pretrained" "custom"
 #modal.state.modelSize="yolov5s.pt"
 modal.state.selectedModel="YOLOv5s"
 modal.state.weightsPath="/yolov5_train/lemons_annotated/5060/weights/best.pt"
-modal.state.device="0"
+modal.state.device="cpu"
 
 
 SERVER_ADDRESS="put your value here"
