+++ supervisely_app/serve/src/modal.html
