+++ supervisely_app/train/src/ui/architectures.html
