+++ supervisely_app/train/src/ui/artifacts.py
