+++ supervisely_app/train/src/ui/classes.html
