+++ supervisely_app/train/src/ui/input_project.py
