+++ supervisely_app/train/src/ui/monitoring.html
