+++ supervisely_app/train/src/ui/splits.html
